Use the optimizer rule `PrepareDeltaScan` to perform the filter optimizations instead of `ActiveOptimisticTransactionRule`.
Consolidate test suites.
