+++ core/src/main/scala/io/delta/sql/DeltaSparkSessionExtension.scala
@@ -95,9 +95,11 @@ class DeltaSparkSessionExtension extends (SparkSessionExtensions => Unit) {
     extensions.injectPostHocResolutionRule { session =>
       new PreprocessTableDelete(session.sessionState.conf)
     }
-    // needs to be in own optimization batch otherwise other optimizer rules will reduce the
-    // generated column partition filters
+    // We don't use `injectOptimizerRule` here as we won't want to apply further optimizations after
+    // `PrepareDeltaScan`.
+    // For example, `ConstantFolding` will break unit tests in `OptimizeGeneratedColumnSuite`.
     extensions.injectPreCBORule { session =>
-      new ActiveOptimisticTransactionRule(session)
+      new PrepareDeltaScan(session)
+    }
   }
 }
