+++ core/src/test/scala/org/apache/spark/sql/delta/perf/OptimizeGeneratedColumnSuite.scala
@@ -16,15 +16,1083 @@
 
 package org.apache.spark.sql.delta.perf
 
-import org.apache.spark.sql.delta.ActiveOptimisticTransactionRule
-import org.apache.spark.sql.types.TimestampType
-import org.apache.spark.util.ThreadUtils
-
+import java.sql.Timestamp
+import java.util.Locale
 import java.util.concurrent.{CountDownLatch, TimeUnit}
 
+import scala.util.matching.Regex
 
-class OptimizeGeneratedColumnSuite extends OptimizeGeneratedColumnSuiteBase {
+// scalastyle:off import.ordering.noEmptyLine
+import org.apache.spark.sql.delta._
+
+import org.apache.spark.sql.delta.sources.DeltaSQLConf.{DELTA_COLLECT_STATS, GENERATED_COLUMN_PARTITION_FILTER_OPTIMIZATION_ENABLED}
+import org.apache.spark.sql.delta.stats.PrepareDeltaScanBase
+
+import org.apache.spark.sql.{DataFrame, Row}
+import org.apache.spark.sql.catalyst.TableIdentifier
+import org.apache.spark.sql.catalyst.expressions.Expression
+import org.apache.spark.sql.execution.{FileSourceScanExec, QueryExecution}
+import org.apache.spark.sql.types.TimestampType
+import org.apache.spark.util.ThreadUtils
+import org.apache.spark.util.Utils
+
+class OptimizeGeneratedColumnSuite extends GeneratedColumnTest {
   import testImplicits._
+
+  private val regex = new Regex(s"(\\S+)\\s(\\S+)\\sGENERATED\\sALWAYS\\sAS\\s\\((.*)\\s?\\)",
+  "col_name", "data_type", "generated_as")
+
+  private def getPushedPartitionFilters(queryExecution: QueryExecution): Seq[Expression] = {
+    queryExecution.executedPlan.collectFirst {
+      case scan: FileSourceScanExec => scan.partitionFilters
+    }.getOrElse(Nil)
+  }
+
+  protected def insertInto(path: String, df: DataFrame) = {
+    df.write.format("delta").mode("append").save(path)
+  }
+
+  /**
+   * Verify we can recognize an `OptimizablePartitionExpression` and generate corresponding
+   * partition filters correctly.
+   *
+   * @param normalColDDL DDL to define a data column
+   * @param generatedPartColDDL a list of generated partition columns defined using the above data
+   *                            column
+   * @param expectedPartitionExpr the expected `OptimizablePartitionExpression` to be recognized
+   * @param filterTestCases test cases for partition filters. The key is the data filter, and the
+   *                        value is the partition filters we should generate.
+   */
+  private def testOptimizablePartitionExpression(
+      schemaString: String,
+      generatedColumns: Map[String, String],
+      expectedPartitionExpr: OptimizablePartitionExpression,
+      auxiliaryTestName: Option[String] = None,
+      filterTestCases: Seq[(String, Seq[String])]): Unit = {
+    test(expectedPartitionExpr.toString + auxiliaryTestName.getOrElse("")) {
+      val normalCol = schemaString.split(" ")(0)
+
+      withTableName("optimizable_partition_expression") { table =>
+        createTable(
+          table,
+          None,
+          schemaString,
+          generatedColumns,
+          generatedColumns.keys.toSeq
+        )
+
+        val metadata = DeltaLog.forTable(spark, TableIdentifier(table)).snapshot.metadata
+        assert(metadata.optimizablePartitionExpressions(normalCol.toLowerCase(Locale.ROOT)) ==
+          expectedPartitionExpr :: Nil)
+        filterTestCases.foreach { filterTestCase =>
+          val partitionFilters = getPushedPartitionFilters(
+            sql(s"SELECT * from $table where ${filterTestCase._1}").queryExecution)
+          assert(partitionFilters.map(_.sql) == filterTestCase._2)
+        }
+      }
+    }
+  }
+
+  /** Format a human readable SQL filter into Spark's compact SQL format */
+  private def compactFilter(filter: String): String = {
+    filter.replaceAllLiterally("\n", "")
+      .replaceAll("(?<=\\)) +(?=\\))", "")
+      .replaceAll("(?<=\\() +(?=\\()", "")
+      .replaceAll("\\) +OR +\\(", ") OR (")
+      .replaceAll("\\) +AND +\\(", ") AND (")
+  }
+
+  testOptimizablePartitionExpression(
+    "eventTime TIMESTAMP, date DATE",
+    Map("date" -> "CAST(eventTime AS DATE)"),
+    expectedPartitionExpr = DatePartitionExpr("date"),
+    auxiliaryTestName = Option(" from cast(timestamp)"),
+    filterTestCases = Seq(
+      "eventTime < '2021-01-01 18:00:00'" ->
+        Seq("((date <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      "eventTime <= '2021-01-01 18:00:00'" ->
+        Seq("((date <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      "eventTime = '2021-01-01 18:00:00'" ->
+        Seq("((date = CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date = CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      "eventTime > '2021-01-01 18:00:00'" ->
+        Seq("((date >= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date >= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      "eventTime >= '2021-01-01 18:00:00'" ->
+        Seq("((date >= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date >= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      "eventTime is null" -> Seq("(date IS NULL)"),
+      // Verify we can reverse the order
+      "'2021-01-01 18:00:00' > eventTime" ->
+        Seq("((date <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      "'2021-01-01 18:00:00' >= eventTime" ->
+        Seq("((date <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      "'2021-01-01 18:00:00' = eventTime" ->
+        Seq("((date = CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date = CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      "'2021-01-01 18:00:00' < eventTime" ->
+        Seq("((date >= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date >= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      "'2021-01-01 18:00:00' <= eventTime" ->
+        Seq("((date >= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+          "OR ((date >= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"),
+      // Verify date type literal. In theory, the best filter should be date < DATE '2021-01-01'.
+      // But Spark's analyzer converts eventTime < '2021-01-01' to
+      // `eventTime` < TIMESTAMP '2021-01-01 00:00:00'. So it's the same as
+      // eventTime < '2021-01-01 18:00:00' for `OptimizeGeneratedColumn`.
+      "eventTime < '2021-01-01'" ->
+        Seq("((date <= CAST(TIMESTAMP '2021-01-01 00:00:00' AS DATE)) " +
+          "OR ((date <= CAST(TIMESTAMP '2021-01-01 00:00:00' AS DATE)) IS NULL))")
+    )
+  )
+
+  testOptimizablePartitionExpression(
+    "eventDate DATE, date DATE",
+    Map("date" -> "CAST(eventDate AS DATE)"),
+    expectedPartitionExpr = DatePartitionExpr("date"),
+    auxiliaryTestName = Option(" from cast(date)"),
+    filterTestCases = Seq(
+      "eventDate < '2021-01-01 18:00:00'" ->
+        Seq("((date <= DATE '2021-01-01') " +
+          "OR ((date <= DATE '2021-01-01') IS NULL))"),
+      "eventDate <= '2021-01-01 18:00:00'" ->
+        Seq("((date <= DATE '2021-01-01') " +
+          "OR ((date <= DATE '2021-01-01') IS NULL))"),
+      "eventDate = '2021-01-01 18:00:00'" ->
+        Seq("((date = DATE '2021-01-01') " +
+          "OR ((date = DATE '2021-01-01') IS NULL))"),
+      "eventDate > '2021-01-01 18:00:00'" ->
+        Seq("((date >= DATE '2021-01-01') " +
+          "OR ((date >= DATE '2021-01-01') IS NULL))"),
+      "eventDate >= '2021-01-01 18:00:00'" ->
+        Seq("((date >= DATE '2021-01-01') " +
+          "OR ((date >= DATE '2021-01-01') IS NULL))"),
+      "eventDate is null" -> Seq("(date IS NULL)"),
+      // Verify we can reverse the order
+      "'2021-01-01 18:00:00' > eventDate" ->
+        Seq("((date <= DATE '2021-01-01') " +
+          "OR ((date <= DATE '2021-01-01') IS NULL))"),
+      "'2021-01-01 18:00:00' >= eventDate" ->
+        Seq("((date <= DATE '2021-01-01') " +
+          "OR ((date <= DATE '2021-01-01') IS NULL))"),
+      "'2021-01-01 18:00:00' = eventDate" ->
+        Seq("((date = DATE '2021-01-01') " +
+          "OR ((date = DATE '2021-01-01') IS NULL))"),
+      "'2021-01-01 18:00:00' < eventDate" ->
+        Seq("((date >= DATE '2021-01-01') " +
+          "OR ((date >= DATE '2021-01-01') IS NULL))"),
+      "'2021-01-01 18:00:00' <= eventDate" ->
+        Seq("((date >= DATE '2021-01-01') " +
+          "OR ((date >= DATE '2021-01-01') IS NULL))"),
+      // Verify date type literal. In theory, the best filter should be date < DATE '2021-01-01'.
+      // But Spark's analyzer converts eventTime < '2021-01-01' to
+      // `eventTime` < TIMESTAMP '2021-01-01 00:00:00'. So it's the same as
+      // eventTime < '2021-01-01 18:00:00' for `OptimizeGeneratedColumn`.
+      "eventDate < '2021-01-01'" ->
+        Seq("((date <= DATE '2021-01-01') " +
+          "OR ((date <= DATE '2021-01-01') IS NULL))")
+    )
+  )
+
+  testOptimizablePartitionExpression(
+    "eventTime TIMESTAMP, year INT, month INT, day INT, hour INT",
+    Map(
+      "year" -> "YEAR(eventTime)",
+      "month" -> "MONTH(eventTime)",
+      "day" -> "DAY(eventTime)",
+      "hour" -> "HOUR(eventTime)"
+    ),
+    expectedPartitionExpr = YearMonthDayHourPartitionExpr("year", "month", "day", "hour"),
+    filterTestCases = Seq(
+      "eventTime < '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (
+            |    (
+            |      (year < year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      OR
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month < month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |    )
+            |    OR
+            |    (
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |      AND
+            |      (day < dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |  )
+            |  OR
+            |  (
+            |    (
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |      AND
+            |      (day = dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |    AND
+            |    (hour <= hour(TIMESTAMP '2021-01-01 18:00:00'))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime <= '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (
+            |    (
+            |      (year < year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      OR
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month < month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |    )
+            |    OR
+            |    (
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |      AND
+            |      (day < dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |  )
+            |  OR
+            |  (
+            |    (
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |      AND
+            |      (day = dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |    AND
+            |    (hour <= hour(TIMESTAMP '2021-01-01 18:00:00'))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime = '2021-01-01 18:00:00'" -> Seq(
+        "(year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))",
+        "(month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))",
+        "(day = dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))",
+        "(hour = hour(TIMESTAMP '2021-01-01 18:00:00'))"
+      ),
+      "eventTime > '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (
+            |    (
+            |      (year > year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      OR
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month > month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |    )
+            |    OR
+            |    (
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |      AND
+            |      (day > dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |  )
+            |  OR
+            |  (
+            |    (
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |      AND
+            |      (day = dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |    AND
+            |    (hour >= hour(TIMESTAMP '2021-01-01 18:00:00'))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime >= '2021-01-01 18:00:00'" ->Seq(
+        compactFilter(
+          """(
+            |  (
+            |    (
+            |      (year > year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      OR
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month > month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |    )
+            |    OR
+            |    (
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |      AND
+            |      (day > dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |  )
+            |  OR
+            |  (
+            |    (
+            |      (
+            |        (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |        AND
+            |        (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      )
+            |      AND
+            |      (day = dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |    AND
+            |    (hour >= hour(TIMESTAMP '2021-01-01 18:00:00'))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime is null" -> Seq(
+        "(year IS NULL)",
+        "(month IS NULL)",
+        "(day IS NULL)",
+        "(hour IS NULL)"
+      )
+    )
+  )
+
+  testOptimizablePartitionExpression(
+    "eventTime TIMESTAMP, year INT, month INT, day INT",
+    Map(
+      "year" -> "YEAR(eventTime)",
+      "month" -> "MONTH(eventTime)",
+      "day" -> "DAY(eventTime)"
+    ),
+    expectedPartitionExpr = YearMonthDayPartitionExpr("year", "month", "day"),
+    filterTestCases = Seq(
+      "eventTime < '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (
+            |    (year < year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    OR
+            |    (
+            |      (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      AND
+            |      (month < month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |  )
+            |  OR
+            |  (
+            |    (
+            |      (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      AND
+            |      (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |    AND
+            |    (day <= dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime <= '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (
+            |    (year < year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    OR
+            |    (
+            |      (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      AND
+            |      (month < month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |  )
+            |  OR
+            |  (
+            |    (
+            |      (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      AND
+            |      (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |    AND
+            |    (day <= dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime = '2021-01-01 18:00:00'" -> Seq(
+        "(year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))",
+        "(month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))",
+        "(day = dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))"
+      ),
+      "eventTime > '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (
+            |    (year > year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    OR
+            |    (
+            |      (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      AND
+            |      (month > month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |  )
+            |  OR
+            |  (
+            |    (
+            |      (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      AND
+            |      (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |    AND
+            |    (day >= dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime >= '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (
+            |    (year > year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    OR
+            |    (
+            |      (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      AND
+            |      (month > month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |  )
+            |  OR
+            |  (
+            |    (
+            |      (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |      AND
+            |      (month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    )
+            |    AND
+            |    (day >= dayofmonth(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime is null" -> Seq(
+        "(year IS NULL)",
+        "(month IS NULL)",
+        "(day IS NULL)"
+      )
+    )
+  )
+
+  testOptimizablePartitionExpression(
+    "eventTime TIMESTAMP, year INT, month INT",
+    // Use different cases to verify we can recognize the same column using different cases in
+    // generation expressions.
+    Map(
+      "year" -> "YEAR(EVENTTIME)",
+      "month" -> "MONTH(eventTime)"
+    ),
+    expectedPartitionExpr = YearMonthPartitionExpr("year", "month"),
+    filterTestCases = Seq(
+      "eventTime < '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (year < year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  OR
+            |  (
+            |    (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    AND
+            |    (month <= month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime <= '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (year < year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  OR
+            |  (
+            |    (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    AND
+            |    (month <= month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime = '2021-01-01 18:00:00'" -> Seq(
+        "(year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))",
+        "(month = month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))"
+      ),
+      "eventTime > '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (year > year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  OR
+            |  (
+            |    (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    AND
+            |    (month >= month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime >= '2021-01-01 18:00:00'" -> Seq(
+        compactFilter(
+          """(
+            |  (year > year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  OR
+            |  (
+            |    (year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |    AND
+            |    (month >= month(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)))
+            |  )
+            |)
+            |""".stripMargin)),
+      "eventTime is null" -> Seq("(year IS NULL)", "(month IS NULL)")
+    )
+  )
+
+  testOptimizablePartitionExpression(
+    "eventTime TIMESTAMP, year INT",
+    Map("year" -> "YEAR(eventTime)"),
+    expectedPartitionExpr = YearPartitionExpr("year"),
+    filterTestCases = Seq(
+      "eventTime < '2021-01-01 18:00:00'" ->
+        Seq("((year <= year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) " +
+          "OR ((year <= year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) IS NULL))"),
+      "eventTime <= '2021-01-01 18:00:00'" ->
+        Seq("((year <= year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) " +
+          "OR ((year <= year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) IS NULL))"),
+      "eventTime = '2021-01-01 18:00:00'" ->
+        Seq("((year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) " +
+          "OR ((year = year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) IS NULL))"),
+      "eventTime > '2021-01-01 18:00:00'" ->
+        Seq("((year >= year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) " +
+          "OR ((year >= year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) IS NULL))"),
+      "eventTime >= '2021-01-01 18:00:00'" ->
+        Seq("((year >= year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) " +
+          "OR ((year >= year(CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE))) IS NULL))"),
+      "eventTime is null" -> Seq("(year IS NULL)")
+    )
+  )
+
+  Seq(("YEAR(eventDate)", " from year(date)"),
+    ("YEAR(CAST(eventDate AS DATE))", " from year(cast(date))"))
+    .foreach { case (partCol, auxTestName) =>
+      testOptimizablePartitionExpression(
+        "eventDate DATE, year INT",
+        Map("year" -> partCol),
+        expectedPartitionExpr = YearPartitionExpr("year"),
+        auxiliaryTestName = Option(auxTestName),
+        filterTestCases = Seq(
+          "eventDate < '2021-01-01'" ->
+            Seq("((year <= year(DATE '2021-01-01')) " +
+              "OR ((year <= year(DATE '2021-01-01')) IS NULL))"),
+          "eventDate <= '2021-01-01'" ->
+            Seq("((year <= year(DATE '2021-01-01')) " +
+              "OR ((year <= year(DATE '2021-01-01')) IS NULL))"),
+          "eventDate = '2021-01-01'" ->
+            Seq("((year = year(DATE '2021-01-01')) " +
+              "OR ((year = year(DATE '2021-01-01')) IS NULL))"),
+          "eventDate > '2021-01-01'" ->
+            Seq("((year >= year(DATE '2021-01-01')) " +
+              "OR ((year >= year(DATE '2021-01-01')) IS NULL))"),
+          "eventDate >= '2021-01-01'" ->
+            Seq("((year >= year(DATE '2021-01-01')) " +
+              "OR ((year >= year(DATE '2021-01-01')) IS NULL))"),
+          "eventDate is null" -> Seq("(year IS NULL)")
+        )
+      )
+    }
+
+  testOptimizablePartitionExpression(
+    "str STRING, substr STRING",
+    Map("substr" -> "SUBSTRING(str, 2, 3)"),
+    expectedPartitionExpr = SubstringPartitionExpr("substr", 2, 3),
+    filterTestCases = Seq(
+      "str < 'foo'" -> Nil,
+      "str <= 'foo'" -> Nil,
+      "str = 'foo'" -> Seq("((substr IS NULL) OR (substr = substring('foo', 2, 3)))"),
+      "str > 'foo'" -> Nil,
+      "str >= 'foo'" -> Nil,
+      "str is null" -> Seq("(substr IS NULL)")
+    )
+  )
+
+  testOptimizablePartitionExpression(
+    "str STRING, substr STRING",
+    Map("substr" -> "SUBSTRING(str, 0, 3)"),
+    expectedPartitionExpr = SubstringPartitionExpr("substr", 0, 3),
+    filterTestCases = Seq(
+      "str < 'foo'" -> Seq("((substr IS NULL) OR (substr <= substring('foo', 0, 3)))"),
+      "str <= 'foo'" -> Seq("((substr IS NULL) OR (substr <= substring('foo', 0, 3)))"),
+      "str = 'foo'" -> Seq("((substr IS NULL) OR (substr = substring('foo', 0, 3)))"),
+      "str > 'foo'" -> Seq("((substr IS NULL) OR (substr >= substring('foo', 0, 3)))"),
+      "str >= 'foo'" -> Seq("((substr IS NULL) OR (substr >= substring('foo', 0, 3)))"),
+      "str is null" -> Seq("(substr IS NULL)")
+    )
+  )
+
+  testOptimizablePartitionExpression(
+    "str STRING, substr STRING",
+    Map("substr" -> "SUBSTRING(str, 1, 3)"),
+    expectedPartitionExpr = SubstringPartitionExpr("substr", 1, 3),
+    filterTestCases = Seq(
+      "str < 'foo'" -> Seq("((substr IS NULL) OR (substr <= substring('foo', 1, 3)))"),
+      "str <= 'foo'" -> Seq("((substr IS NULL) OR (substr <= substring('foo', 1, 3)))"),
+      "str = 'foo'" -> Seq("((substr IS NULL) OR (substr = substring('foo', 1, 3)))"),
+      "str > 'foo'" -> Seq("((substr IS NULL) OR (substr >= substring('foo', 1, 3)))"),
+      "str >= 'foo'" -> Seq("((substr IS NULL) OR (substr >= substring('foo', 1, 3)))"),
+      "str is null" -> Seq("(substr IS NULL)")
+    )
+  )
+
+  test("end-to-end optimizable partition expression") {
+    withTempDir { tempDir =>
+      withTableName("optimizable_partition_expression") { table =>
+
+        createTable(
+          table,
+          Some(tempDir.getCanonicalPath),
+          "c1 INT, c2 TIMESTAMP, c3 DATE",
+          Map("c3" -> "CAST(c2 AS DATE)"),
+          Seq("c3")
+        )
+
+          Seq(
+            Tuple2(1, "2020-12-31 11:00:00"),
+            Tuple2(2, "2021-01-01 12:00:00"),
+            Tuple2(3, "2021-01-02 13:00:00")
+          ).foreach { values =>
+            insertInto(
+              tempDir.getCanonicalPath,
+              Seq(values).toDF("c1", "c2")
+                .withColumn("c2", $"c2".cast(TimestampType))
+            )
+          }
+        assert(tempDir.listFiles().map(_.getName).toSet ==
+          Set("c3=2021-01-01", "c3=2021-01-02", "c3=2020-12-31", "_delta_log"))
+        // Delete folders which should not be read if we generate the partition filters correctly
+        tempDir.listFiles().foreach { f =>
+          if (f.getName != "c3=2021-01-01" && f.getName != "_delta_log") {
+            Utils.deleteRecursively(f)
+          }
+        }
+        assert(tempDir.listFiles().map(_.getName).toSet == Set("c3=2021-01-01", "_delta_log"))
+        checkAnswer(
+          sql(s"select * from $table where " +
+            s"c2 >= '2021-01-01 12:00:00' AND c2 <= '2021-01-01 18:00:00'"),
+          Row(2, sqlTimestamp("2021-01-01 12:00:00"), sqlDate("2021-01-01")))
+        // Verify `OptimizeGeneratedColumn` doesn't mess up Projects.
+        checkAnswer(
+          sql(s"select c1 from $table where " +
+            s"c2 >= '2021-01-01 12:00:00' AND c2 <= '2021-01-01 18:00:00'"),
+          Row(2))
+
+        // Check both projection orders to make sure projection orders are handled correctly
+        checkAnswer(
+          sql(s"select c1, c2 from $table where " +
+            s"c2 >= '2021-01-01 12:00:00' AND c2 <= '2021-01-01 18:00:00'"),
+          Row(2, Timestamp.valueOf("2021-01-01 12:00:00")))
+        checkAnswer(
+          sql(s"select c2, c1 from $table where " +
+            s"c2 >= '2021-01-01 12:00:00' AND c2 <= '2021-01-01 18:00:00'"),
+          Row(Timestamp.valueOf("2021-01-01 12:00:00"), 2))
+
+        // Verify the optimization works for limit.
+        val limitQuery = sql(
+          s"""select * from $table
+             |where c2 >= '2021-01-01 12:00:00' AND c2 <= '2021-01-01 18:00:00'
+             |limit 10""".stripMargin)
+        val expectedPartitionFilters = Seq(
+          "((c3 >= CAST(TIMESTAMP '2021-01-01 12:00:00' AS DATE)) " +
+            "OR ((c3 >= CAST(TIMESTAMP '2021-01-01 12:00:00' AS DATE)) IS NULL))",
+          "((c3 <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) " +
+            "OR ((c3 <= CAST(TIMESTAMP '2021-01-01 18:00:00' AS DATE)) IS NULL))"
+        )
+        assert(expectedPartitionFilters ==
+          getPushedPartitionFilters(limitQuery.queryExecution).map(_.sql))
+        checkAnswer(limitQuery, Row(2, sqlTimestamp("2021-01-01 12:00:00"), sqlDate("2021-01-01")))
+      }
+    }
+  }
+
+  test("empty string and null ambiguity in a partition column") {
+    withTempDir { tempDir =>
+      withTableName("optimizable_partition_expression") { table =>
+        createTable(
+          table,
+          Some(tempDir.getCanonicalPath),
+          "c1 STRING, c2 STRING",
+          Map("c2" -> "SUBSTRING(c1, 1, 4)"),
+          Seq("c2")
+        )
+        insertInto(
+          tempDir.getCanonicalPath,
+          Seq(Tuple1("")).toDF("c1")
+        )
+        checkAnswer(
+          sql(s"select * from $table where c1 = ''"),
+          Row("", null))
+        // The following check shows the weird behavior of SPARK-24438 and confirms the generated
+        // partition filter doesn't impact the answer.
+        withSQLConf(GENERATED_COLUMN_PARTITION_FILTER_OPTIMIZATION_ENABLED.key -> "false") {
+          checkAnswer(
+            sql(s"select * from $table where c1 = ''"),
+            Row("", null))
+        }
+      }
+    }
+  }
+
+  test("substring on multibyte characters") {
+    withTempDir { tempDir =>
+      withTableName("multibyte_characters") { table =>
+        createTable(
+          table,
+          Some(tempDir.getCanonicalPath),
+          "c1 STRING, c2 STRING",
+          Map("c2" -> "SUBSTRING(c1, 1, 2)"),
+          Seq("c2")
+        )
+        // scalastyle:off nonascii
+        insertInto(
+          tempDir.getCanonicalPath,
+          Seq(Tuple1("一二三四")).toDF("c1")
+        )
+        val testQuery = s"select * from $table where c1 > 'abcd'"
+        assert("((c2 IS NULL) OR (c2 >= substring('abcd', 1, 2)))" :: Nil ==
+          getPushedPartitionFilters(sql(testQuery).queryExecution).map(_.sql))
+        checkAnswer(
+          sql(testQuery),
+          Row("一二三四", "一二"))
+        // scalastyle:on nonascii
+      }
+    }
+  }
+
+  testOptimizablePartitionExpression(
+    "eventTime TIMESTAMP, month STRING",
+    Map("month" -> "DATE_FORMAT(eventTime, 'yyyy-MM')"),
+    expectedPartitionExpr = DateFormatPartitionExpr("month", "yyyy-MM"),
+    auxiliaryTestName = Option(" from timestamp"),
+    filterTestCases = Seq(
+      "eventTime < '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') <= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') <= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), " +
+          "'yyyy-MM')) IS NULL))"),
+      "eventTime <= '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') <= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') <= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), " +
+          "'yyyy-MM')) IS NULL))"),
+      "eventTime = '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') = " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') = " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), " +
+          "'yyyy-MM')) IS NULL))"),
+      "eventTime > '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') >= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') >= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), " +
+          "'yyyy-MM')) IS NULL))"),
+      "eventTime >= '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') >= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') >= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM'), " +
+          "'yyyy-MM')) IS NULL))"),
+      "eventTime is null" -> Seq("(month IS NULL)")
+    )
+  )
+
+  testOptimizablePartitionExpression(
+    "eventDate DATE, month STRING",
+    Map("month" -> "DATE_FORMAT(eventDate, 'yyyy-MM')"),
+    expectedPartitionExpr = DateFormatPartitionExpr("month", "yyyy-MM"),
+    auxiliaryTestName = Option(" from cast(date)"),
+    filterTestCases = Seq(
+      "eventDate < '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') <= unix_timestamp(date_format(CAST(" +
+          "DATE '2021-06-28' AS TIMESTAMP), 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') <= " +
+          "unix_timestamp(date_format(CAST(DATE '2021-06-28' AS TIMESTAMP), " +
+          "'yyyy-MM'), 'yyyy-MM')) IS NULL))"),
+      "eventDate <= '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') <= unix_timestamp(date_format(CAST(" +
+          "DATE '2021-06-28' AS TIMESTAMP), 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') <= " +
+          "unix_timestamp(date_format(CAST(DATE '2021-06-28' AS TIMESTAMP), " +
+          "'yyyy-MM'), 'yyyy-MM')) IS NULL))"),
+      "eventDate = '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') = unix_timestamp(date_format(CAST(" +
+          "DATE '2021-06-28' AS TIMESTAMP), 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') = " +
+          "unix_timestamp(date_format(CAST(DATE '2021-06-28' AS TIMESTAMP), " +
+          "'yyyy-MM'), 'yyyy-MM')) IS NULL))"),
+      "eventDate > '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') >= unix_timestamp(date_format(CAST(" +
+          "DATE '2021-06-28' AS TIMESTAMP), 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') >= " +
+          "unix_timestamp(date_format(CAST(DATE '2021-06-28' AS TIMESTAMP), " +
+          "'yyyy-MM'), 'yyyy-MM')) IS NULL))"),
+      "eventDate >= '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(month, 'yyyy-MM') >= unix_timestamp(date_format(CAST(" +
+          "DATE '2021-06-28' AS TIMESTAMP), 'yyyy-MM'), 'yyyy-MM')) " +
+          "OR ((unix_timestamp(month, 'yyyy-MM') >= " +
+          "unix_timestamp(date_format(CAST(DATE '2021-06-28' AS TIMESTAMP), " +
+          "'yyyy-MM'), 'yyyy-MM')) IS NULL))"),
+      "eventDate is null" -> Seq("(month IS NULL)")
+    )
+  )
+
+  testOptimizablePartitionExpression(
+    "eventTime TIMESTAMP, hour STRING",
+    Map("hour" -> "(DATE_FORMAT(eventTime, 'yyyy-MM-dd-HH'))"),
+    expectedPartitionExpr = DateFormatPartitionExpr("hour", "yyyy-MM-dd-HH"),
+    filterTestCases = Seq(
+      "eventTime < '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(hour, 'yyyy-MM-dd-HH') <= unix_timestamp(" +
+          "date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) " +
+          "OR ((unix_timestamp(hour, 'yyyy-MM-dd-HH') <= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', " +
+          "'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) IS NULL))"),
+      "eventTime <= '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(hour, 'yyyy-MM-dd-HH') <= unix_timestamp(" +
+          "date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) " +
+          "OR ((unix_timestamp(hour, 'yyyy-MM-dd-HH') <= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', " +
+          "'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) IS NULL))"),
+      "eventTime = '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(hour, 'yyyy-MM-dd-HH') = unix_timestamp(" +
+          "date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) " +
+          "OR ((unix_timestamp(hour, 'yyyy-MM-dd-HH') = " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', " +
+          "'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) IS NULL))"),
+      "eventTime > '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(hour, 'yyyy-MM-dd-HH') >= unix_timestamp(" +
+          "date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) " +
+          "OR ((unix_timestamp(hour, 'yyyy-MM-dd-HH') >= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', " +
+          "'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) IS NULL))"),
+      "eventTime >= '2021-06-28 18:00:00'" ->
+        Seq("((unix_timestamp(hour, 'yyyy-MM-dd-HH') >= unix_timestamp(" +
+          "date_format(TIMESTAMP '2021-06-28 18:00:00', 'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) " +
+          "OR ((unix_timestamp(hour, 'yyyy-MM-dd-HH') >= " +
+          "unix_timestamp(date_format(TIMESTAMP '2021-06-28 18:00:00', " +
+          "'yyyy-MM-dd-HH'), 'yyyy-MM-dd-HH')) IS NULL))"),
+      "eventTime is null" -> Seq("(hour IS NULL)")
+    )
+  )
+
+  test("five digits year in a year month day partition column") {
+    withTempDir { tempDir =>
+      withTableName("optimizable_partition_expression") { table =>
+        createTable(
+          table,
+          Some(tempDir.getCanonicalPath),
+          "c1 TIMESTAMP, c2 INT, c3 INT, c4 INT",
+          Map(
+            "c2" -> "YEAR(c1)",
+            "c3" -> "MONTH(c1)",
+            "c4" -> "DAY(c1)"
+          ),
+          Seq("c2", "c3", "c4")
+        )
+        insertInto(
+          tempDir.getCanonicalPath,
+          Seq(Tuple1("12345-07-15 18:00:00"))
+            .toDF("c1")
+            .withColumn("c1", $"c1".cast(TimestampType))
+        )
+
+        checkAnswer(
+          sql(s"select * from $table where c1 = CAST('12345-07-15 18:00:00' as timestamp)"),
+          Row(new Timestamp(327420320400000L), 12345, 7, 15))
+        withSQLConf(GENERATED_COLUMN_PARTITION_FILTER_OPTIMIZATION_ENABLED.key -> "false") {
+          checkAnswer(
+            sql(s"select * from $table where c1 = CAST('12345-07-15 18:00:00' as timestamp)"),
+            Row(new Timestamp(327420320400000L), 12345, 7, 15))
+        }
+      }
+    }
+  }
+
+  test("five digits year in a date_format yyyy-MM partition column") {
+    withTempDir { tempDir =>
+      withTableName("optimizable_partition_expression") { table =>
+        createTable(
+          table,
+          Some(tempDir.getCanonicalPath),
+          "c1 TIMESTAMP, c2 STRING",
+          Map("c2" -> "DATE_FORMAT(c1, 'yyyy-MM')"),
+          Seq("c2")
+        )
+        insertInto(
+          tempDir.getCanonicalPath,
+          Seq(Tuple1("12345-07-15 18:00:00"))
+            .toDF("c1")
+            .withColumn("c1", $"c1".cast(TimestampType))
+        )
+
+        checkAnswer(
+          sql(s"select * from $table where c1 = CAST('12345-07-15 18:00:00' as timestamp)"),
+          Row(new Timestamp(327420320400000L), "+12345-07"))
+        withSQLConf(GENERATED_COLUMN_PARTITION_FILTER_OPTIMIZATION_ENABLED.key -> "false") {
+          checkAnswer(
+            sql(s"select * from $table where c1 = CAST('12345-07-15 18:00:00' as timestamp)"),
+            Row(new Timestamp(327420320400000L), "+12345-07"))
+        }
+      }
+    }
+  }
+
+  test("five digits year in a date_format yyyy-MM-dd-HH partition column") {
+    withTempDir { tempDir =>
+      withTableName("optimizable_partition_expression") { table =>
+        createTable(
+          table,
+          Some(tempDir.getCanonicalPath),
+          "c1 TIMESTAMP, c2 STRING",
+          Map("c2" -> "DATE_FORMAT(c1, 'yyyy-MM-dd-HH')"),
+          Seq("c2")
+        )
+        insertInto(
+          tempDir.getCanonicalPath,
+          Seq(Tuple1("12345-07-15 18:00:00"))
+            .toDF("c1")
+            .withColumn("c1", $"c1".cast(TimestampType))
+        )
+
+        checkAnswer(
+          sql(s"select * from $table where c1 = CAST('12345-07-15 18:00:00' as timestamp)"),
+          Row(new Timestamp(327420320400000L), "+12345-07-15-18"))
+        withSQLConf(GENERATED_COLUMN_PARTITION_FILTER_OPTIMIZATION_ENABLED.key -> "false") {
+          checkAnswer(
+            sql(s"select * from $table where c1 = CAST('12345-07-15 18:00:00' as timestamp)"),
+            Row(new Timestamp(327420320400000L), "+12345-07-15-18"))
+        }
+      }
+    }
+  }
+
+  test("end-to-end test of behaviors of write/read null on partition column") {
+    //              unix_timestamp('12345-12', 'yyyy-MM') | unix_timestamp('+12345-12', 'yyyy-MM')
+    //  EXCEPTION               fail                     |           327432240000
+    //  CORRECTED               null                     |           327432240000
+    //  LEGACY               327432240000                |               null
+    withTempDir { tempDir =>
+      withTableName("optimizable_partition_expression") { table =>
+        createTable(
+          table,
+          Some(tempDir.getCanonicalPath),
+          "c1 TIMESTAMP, c2 STRING",
+          Map("c2" -> "DATE_FORMAT(c1, 'yyyy-MM')"),
+          Seq("c2")
+        )
+
+        // write in LEGACY
+        withSQLConf(
+        "spark.sql.legacy.timeParserPolicy" -> "CORRECTED"
+        ) {
+          insertInto(
+            tempDir.getCanonicalPath,
+            Seq(Tuple1("12345-07-01 00:00:00"))
+              .toDF("c1")
+              .withColumn("c1", $"c1".cast(TimestampType))
+          )
+          insertInto(
+            tempDir.getCanonicalPath,
+            Seq(Tuple1("+23456-07-20 18:30:00"))
+              .toDF("c1")
+              .withColumn("c1", $"c1".cast(TimestampType))
+          )
+        }
+
+        // write in LEGACY
+        withSQLConf(
+          "spark.sql.legacy.timeParserPolicy" -> "LEGACY"
+        ) {
+          insertInto(
+            tempDir.getCanonicalPath,
+            Seq(Tuple1("+12349-07-01 00:00:00"))
+              .toDF("c1")
+              .withColumn("c1", $"c1".cast(TimestampType))
+          )
+          insertInto(
+            tempDir.getCanonicalPath,
+            Seq(Tuple1("+30000-12-30 20:00:00"))
+              .toDF("c1")
+              .withColumn("c1", $"c1".cast(TimestampType))
+          )
+        }
+
+        // we have partitions based on CORRECTED + LEGACY parser (with +)
+        assert(tempDir.listFiles().map(_.getName).toSet ==
+          Set("c2=+23456-07", "c2=12349-07", "c2=30000-12", "c2=+12345-07", "_delta_log"))
+
+        // read behaviors in CORRECTED, we still can query correctly
+        withSQLConf("spark.sql.legacy.timeParserPolicy" -> "CORRECTED") {
+          checkAnswer(
+            sql(s"select (unix_timestamp('+20000-01', 'yyyy-MM')) as value"),
+            Row(568971849600L)
+          )
+          checkAnswer(
+            sql(s"select (unix_timestamp('20000-01', 'yyyy-MM')) as value"),
+            Row(null)
+          )
+          checkAnswer(
+            sql(s"select * from $table where " +
+              s"c1 >= '20000-01-01 12:00:00'"),
+            // 23456-07-20 18:30:00
+            Row(new Timestamp(678050098200000L), "+23456-07") ::
+              // 30000-12-30 20:00:00
+              Row(new Timestamp(884572891200000L), "30000-12") :: Nil
+          )
+        }
+
+        // read behaviors in LEGACY, we still can query correctly
+        withSQLConf("spark.sql.legacy.timeParserPolicy" -> "LEGACY") {
+          checkAnswer(
+            sql(s"select (unix_timestamp('+20000-01', 'yyyy-MM')) as value"),
+            Row(null)
+          )
+          checkAnswer(
+            sql(s"select (unix_timestamp('20000-01', 'yyyy-MM')) as value"),
+            Row(568971849600L)
+          )
+          checkAnswer(
+            sql(s"select * from $table where " +
+              s"c1 >= '20000-01-01 12:00:00'"),
+            // 23456-07-20 18:30:00
+            Row(new Timestamp(678050098200000L), "+23456-07") ::
+              // 30000-12-30 20:00:00
+              Row(new Timestamp(884572891200000L), "30000-12") :: Nil
+          )
+        }
+      }
+    }
+  }
 
   test("generated partition filters should avoid conflicts") {
     withTempDir { tempDir =>
@@ -47,7 +1115,7 @@ class OptimizeGeneratedColumnSuite extends OptimizeGeneratedColumnSuiteBase {
         val unblockQueries = new CountDownLatch(1)
         val waitForAllQueries = new CountDownLatch(2)
 
-        ActiveOptimisticTransactionRule.withCallbackOnGetSnapshot(_ => {
+        PrepareDeltaScanBase.withCallbackOnGetDeltaScanGenerator(_ => {
           waitForAllQueries.countDown()
           assert(
             unblockQueries.await(30, TimeUnit.SECONDS),
